fix(profile): fix DatePickerDialog missing content param + ChatScreen Float->Int
PRE-EXISTING BUILD FIXES
========================
1. ProfileScreen DatePickerDialog was missing the required 'content' trailing
   lambda (DatePicker composable). In Material3 1.3.0 (Compose BOM 2024.09.00),
   DatePickerDialog requires a content lambda that renders the DatePicker itself.
   Added: ) { DatePicker(state = datePickerState) }

2. ChatScreen ChatVoiceRecordingBar durationSeconds parameter expects Int but
   recordingDurationSec is Float (MutableStateFlow<Float>). Fixed with .toInt().

Both were pre-existing compile errors that blocked the entire app from building.

Profile fix: also removed supportingText from OutlinedTextField (replaced with
a manual Text below the field) to keep the dialog simpler.

Build now passes: ./gradlew :app:compileDebugKotlin BUILD SUCCESSFUL

+++ app/src/main/java/com/example/ui/screens/ChatScreen.kt
@@ -529,7 +529,7 @@ fun ChatScreen(
                     // Bottom Input or Voice Recording Bar
                     if (isRecordingVoice) {
                         ChatVoiceRecordingBar(
-                            durationSeconds = recordingDurationSec,
+                            durationSeconds = recordingDurationSec.toInt(),
                             amplitudes = recordingAmplitudes,
                             onCancel = { viewModel.cancelVoiceRecording() },
                             onSend = { viewModel.sendVoiceMessage() }
